+++ HISTORY.rst
@@ -10,6 +10,7 @@ History
     * Add examples and implementation notes to each metric  
     * Update listing in 'details_for_all_metrics.py'
     * Update the 'Why jsmetrics' section with notes about what jets are and quick start help
+    * Add some Examples of Use for each type of metric 
 * Add ws_threshold parameter to Pena-Ortiz method
 * Rename Pena-Ortiz method to 'pena_ortiz_et_al_2016' (as earlier paper)